agents/requirements.txt
# ADK Agent Dependencies
google-adk>=1.11.0
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0
aiohttp>=3.8.0
python-dotenv>=1.0.0

# Enhanced System Dependencies
requests>=2.31.0
dataclasses>=0.6; python_version<"3.7"

# Development dependencies
pytest>=7.0.0
black>=23.0.0
flake8>=6.0.0
